# Flatten `chain`s optimization