Baseline (best result from https://github.com/gcanti/io-ts/pull/134)

```
space-object (good) x 514,274 ops/sec ±0.45% (88 runs sampled)
space-object (bad) x 479,642 ops/sec ±0.54% (86 runs sampled)
```

After this change

```
space-object (good) x 592,606 ops/sec ±0.52% (87 runs sampled)
space-object (bad) x 532,306 ops/sec ±1.18% (81 runs sampled)
```